pyproject.toml
[tool.poetry]
name = "gcp-workshop-lab"
version = "0.1.0"
description = ""
authors = ["Viktor Shytiuk <vshyt@softserveinc.com>"]

[tool.poetry.dependencies]
python = "~3.7"
Faker = "6.5.0"
google-cloud-storage = "1.36.1"
google-cloud-bigquery = "2.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
